Refactor prayerLinking
